Revert "Modify knitro solver factory structure to fit into ac solver factory structure"
This reverts commit baf96411e71ac6a04ac96ae2f58795096310ab6d.
