+++ src/main/java/com/powsybl/openloadflow/ac/solver/KnitroSolverFactory.java
@@ -1,0 +1,31 @@
+/**
+ * Copyright (c) 2024, Artelys (https://www.artelys.com/)
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ * SPDX-License-Identifier: MPL-2.0
+ */
+package com.powsybl.openloadflow.ac.solver;
+
+import com.powsybl.openloadflow.ac.AcLoadFlowParameters;
+import com.powsybl.openloadflow.ac.equations.AcEquationType;
+import com.powsybl.openloadflow.ac.equations.AcVariableType;
+import com.powsybl.openloadflow.equations.EquationSystem;
+import com.powsybl.openloadflow.equations.EquationVector;
+import com.powsybl.openloadflow.equations.JacobianMatrix;
+import com.powsybl.openloadflow.equations.TargetVector;
+import com.powsybl.openloadflow.network.LfNetwork;
+
+/**
+ * @author Pierre Arvy {@literal <pierre.arvy at artelys.com>}
+ */
+public class KnitroSolverFactory implements AcSolverFactory {
+
+    @Override
+    public AcSolver create(LfNetwork network, AcLoadFlowParameters parameters, EquationSystem<AcVariableType, AcEquationType> equationSystem,
+                           JacobianMatrix<AcVariableType, AcEquationType> j, TargetVector<AcVariableType, AcEquationType> targetVector,
+                           EquationVector<AcVariableType, AcEquationType> equationVector) {
+        return new KnitroSolver(network, parameters.getKnitroSolverParameters(), equationSystem,
+                j, targetVector, equationVector, parameters.isDetailedReport());
+    }
+}
