+++ src/main/java/com/powsybl/openloadflow/ac/solver/KnitroSolverParameters.java
@@ -5,10 +5,9 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  * SPDX-License-Identifier: MPL-2.0
  */
-package com.powsybl;
+package com.powsybl.openloadflow.ac.solver;
 
 import com.powsybl.openloadflow.OpenLoadFlowParameters;
-import com.powsybl.openloadflow.ac.solver.*;
 
 import java.util.Objects;
 
@@ -17,29 +16,16 @@ import java.util.Objects;
  * @author Jeanne Archambault {@literal <jeanne.archambault at artelys.com>}
  */
 
-public class KnitroSolverParameters implements AcSolverParameters {
+public class KnitroSolverParameters {
 
     public static final int DEFAULT_GRADIENT_COMPUTATION_MODE = 1; // Specifies how the Jacobian matrix is computed
     public static final int DEFAULT_GRADIENT_USER_ROUTINE = 2; // If the user chooses to pass the exact Jacobian to knitro, specifies the sparsity pattern for the Jacobian matrix.
     public static final double DEFAULT_LOWER_VOLTAGE_BOUND = 0.5; // Lower bound for voltage magnitude
     public static final double DEFAULT_UPPER_VOLTAGE_BOUND = 1.5; // Upper bound for voltage magnitude
     public static final int DEFAULT_MAX_ITERATIONS = 200;
-    public static final StateVectorScalingMode DEFAULT_STATE_VECTOR_SCALING_MODE = StateVectorScalingMode.NONE;
-    public static final boolean ALWAYS_UPDATE_NETWORK_DEFAULT_VALUE = false;
 
-//    public KnitroSolverParameters() {
-//        super(DEFAULT_MAX_ITERATIONS);
-//    }
-
-    private StateVectorScalingMode stateVectorScalingMode = DEFAULT_STATE_VECTOR_SCALING_MODE;
-
-    private int lineSearchStateVectorScalingMaxIteration = LineSearchStateVectorScaling.DEFAULT_MAX_ITERATION;
-
-    private double lineSearchStateVectorScalingStepFold = LineSearchStateVectorScaling.DEFAULT_STEP_FOLD;
-
-    private double maxVoltageChangeStateVectorScalingMaxDv = MaxVoltageChangeStateVectorScaling.DEFAULT_MAX_DV;
-
-    private double maxVoltageChangeStateVectorScalingMaxDphi = MaxVoltageChangeStateVectorScaling.DEFAULT_MAX_DPHI;
+    public KnitroSolverParameters() {
+    }
 
     private int gradientComputationMode = DEFAULT_GRADIENT_COMPUTATION_MODE;
 
@@ -51,7 +37,7 @@ public class KnitroSolverParameters implements AcSolverParameters {
 
     private KnitroSolverStoppingCriteria stoppingCriteria = new DefaultKnitroSolverStoppingCriteria();
 
-    private boolean alwaysUpdateNetwork = ALWAYS_UPDATE_NETWORK_DEFAULT_VALUE;
+    private boolean alwaysUpdateNetwork = OpenLoadFlowParameters.ALWAYS_UPDATE_NETWORK_DEFAULT_VALUE;
 
     private int maxIterations = DEFAULT_MAX_ITERATIONS;
 
@@ -115,58 +101,12 @@ public class KnitroSolverParameters implements AcSolverParameters {
         return this;
     }
 
-    public StateVectorScalingMode getStateVectorScalingMode() {
-        return stateVectorScalingMode;
-    }
-
-    public KnitroSolverParameters setStateVectorScalingMode(StateVectorScalingMode stateVectorScalingMode) {
-        this.stateVectorScalingMode = Objects.requireNonNull(stateVectorScalingMode);
-        return this;
-    }
-
     public boolean isAlwaysUpdateNetwork() {
         return alwaysUpdateNetwork;
     }
 
     public KnitroSolverParameters setAlwaysUpdateNetwork(boolean alwaysUpdateNetwork) {
         this.alwaysUpdateNetwork = alwaysUpdateNetwork;
-        return this;
-    }
-
-    public int getLineSearchStateVectorScalingMaxIteration() {
-        return lineSearchStateVectorScalingMaxIteration;
-    }
-
-    public KnitroSolverParameters setLineSearchStateVectorScalingMaxIteration(int lineSearchStateVectorScalingMaxIteration) {
-        this.lineSearchStateVectorScalingMaxIteration = lineSearchStateVectorScalingMaxIteration;
-        return this;
-
-    }
-
-    public double getLineSearchStateVectorScalingStepFold() {
-        return lineSearchStateVectorScalingStepFold;
-    }
-
-    public KnitroSolverParameters setLineSearchStateVectorScalingStepFold(double lineSearchStateVectorScalingStepFold) {
-        this.lineSearchStateVectorScalingStepFold = lineSearchStateVectorScalingStepFold;
-        return this;
-    }
-
-    public double getMaxVoltageChangeStateVectorScalingMaxDv() {
-        return maxVoltageChangeStateVectorScalingMaxDv;
-    }
-
-    public KnitroSolverParameters setMaxVoltageChangeStateVectorScalingMaxDv(double maxVoltageChangeStateVectorScalingMaxDv) {
-        this.maxVoltageChangeStateVectorScalingMaxDv = maxVoltageChangeStateVectorScalingMaxDv;
-        return this;
-    }
-
-    public double getMaxVoltageChangeStateVectorScalingMaxDphi() {
-        return maxVoltageChangeStateVectorScalingMaxDphi;
-    }
-
-    public KnitroSolverParameters setMaxVoltageChangeStateVectorScalingMaxDphi(double maxVoltageChangeStateVectorScalingMaxDphi) {
-        this.maxVoltageChangeStateVectorScalingMaxDphi = maxVoltageChangeStateVectorScalingMaxDphi;
         return this;
     }
