+++ src/main/java/com/powsybl/openloadflow/ac/solver/KnitroSolverStoppingCriteria.java
@@ -5,7 +5,7 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  * SPDX-License-Identifier: MPL-2.0
  */
-package com.powsybl;
+package com.powsybl.openloadflow.ac.solver;
 
 /**
  * @author Jeanne Archambault {@literal <jeanne.archambault at artelys.com>}
