+++ src/main/java/com/powsybl/openloadflow/ac/solver/NonLinearExternalSolverUtils.java
@@ -6,7 +6,7 @@
  * SPDX-License-Identifier: MPL-2.0
  */
 
-package com.powsybl;
+package com.powsybl.openloadflow.ac.solver;
 import com.powsybl.openloadflow.ac.equations.AcVariableType;
 import com.powsybl.openloadflow.ac.equations.AcEquationType;
 import com.powsybl.openloadflow.equations.EquationTerm;
